+++ anyblok/tests/test_model.py
@@ -262,7 +262,7 @@ class TestModel2(DBTestCase):
             class Test:
 
                 @classmethod
-                def define_table_args(cls, table_args, properties):
+                def define_table_args(cls):
                     return (val,)
 
         registry = self.init_registry(add_in_registry)
@@ -279,15 +279,15 @@ class TestModel2(DBTestCase):
             class Test:
 
                 @classmethod
-                def define_table_args(cls, table_args, properties):
+                def define_table_args(cls):
                     return (val,)
 
             @register(Model)  # noqa
             class Test:
 
                 @classmethod
-                def define_table_args(cls, table_args, properties):
-                    return table_args + (val2,)
+                def define_table_args(cls):
+                    return super(Test, cls).define_table_args() + (val2,)
 
         registry = self.init_registry(add_in_registry)
         self.assertEqual(len(registry.Test.__table_args__), 2)
@@ -314,7 +314,7 @@ class TestModel2(DBTestCase):
             class Test:
 
                 @classmethod
-                def define_mapper_args(cls, mapper_args, properties):
+                def define_mapper_args(cls):
                     return {val: val}
 
         registry = self.init_registry(add_in_registry)
@@ -331,14 +331,15 @@ class TestModel2(DBTestCase):
             class Test:
 
                 @classmethod
-                def define_mapper_args(cls, mapper_args, properties):
+                def define_mapper_args(cls):
                     return {val: val}
 
             @register(Model)  # noqa
             class Test:
 
                 @classmethod
-                def define_mapper_args(cls, mapper_args, properties):
+                def define_mapper_args(cls):
+                    mapper_args = super(Test, cls).define_mapper_args()
                     mapper_args.update({val2: val2})
                     return mapper_args